[Split CSSPropertyParserHelpers] Part 2: String, URL, Position and Image https://bugs.webkit.org/show_bug.cgi?id=275780
Reviewed by Darin Adler.

Further split up CSSPropertyParserHelpers into type specific files.

* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
    - Add new files.

* Source/WebCore/css/parser/CSSPropertyParserConsumer+Ident.cpp:
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Ident.h:
    - Move custom and dashed ident consumers.

* Source/WebCore/css/parser/CSSPropertyParserConsumer+Image.cpp: Copied from Source/WebCore/css/parser/CSSPropertyParserHelpers.cpp.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Image.h: Copied from Source/WebCore/css/parser/CSSPropertyParserHelpers.h.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Position.cpp: Copied from Source/WebCore/css/parser/CSSPropertyParserHelpers.cpp.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Position.h: Copied from Source/WebCore/css/parser/CSSPropertyParserHelpers.h.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+String.cpp: Copied from Source/WebCore/css/parser/CSSPropertyParserHelpers.cpp.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+String.h: Copied from Source/WebCore/css/parser/CSSPropertyParserHelpers.h.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+URL.cpp: Copied from Source/WebCore/css/parser/CSSPropertyParserHelpers.cpp.
* Source/WebCore/css/parser/CSSPropertyParserConsumer+URL.h: Copied from Source/WebCore/css/parser/CSSPropertyParserHelpers.h.
    - Move consume functions to new files.

* Source/WebCore/css/parser/CSSPropertyParserHelpers.cpp:
* Source/WebCore/css/parser/CSSPropertyParserHelpers.h:
    - Remove moved functions.

* Source/WebCore/css/parser/CSSPropertyParser.cpp:
* Source/WebCore/css/parser/CSSPropertyParserWorkerSafe.cpp:
* Source/WebCore/css/process-css-properties.py:
    - Update includes.

Canonical link: https://commits.webkit.org/280294@main
