+++ Source/WebCore/css/parser/CSSPropertyParserConsumer+URL.cpp
@@ -1,0 +1,71 @@
+/*
+ * Copyright (C) 2024 Samuel Weinig <sam@webkit.org>
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the
+ *    documentation and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY APPLE INC. ``AS IS'' AND ANY
+ * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+ * PURPOSE ARE DISCLAIMED.  IN NO EVENT SHALL APPLE INC. OR
+ * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+ * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+ * PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
+ * OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#include "config.h"
+#include "CSSPropertyParserConsumer+URL.h"
+
+#include "CSSParserTokenRange.h"
+#include "CSSPrimitiveValue.h"
+#include "CSSValueKeywords.h"
+#include <wtf/text/StringView.h>
+
+namespace WebCore {
+namespace CSSPropertyParserHelpers {
+
+// MARK: <url>
+// https://drafts.csswg.org/css-values/#urls
+
+StringView consumeURLRaw(CSSParserTokenRange& range)
+{
+    auto& token = range.peek();
+    if (token.type() == UrlToken) {
+        range.consumeIncludingWhitespace();
+        return token.value();
+    }
+
+    if (token.functionId() == CSSValueUrl) {
+        auto rangeCopy = range;
+        auto args = rangeCopy.consumeBlock();
+        auto& next = args.consumeIncludingWhitespace();
+        if (next.type() == BadStringToken || !args.atEnd())
+            return StringView();
+        ASSERT(next.type() == StringToken);
+        range = rangeCopy;
+        range.consumeWhitespace();
+        return next.value();
+    }
+
+    return { };
+}
+
+RefPtr<CSSPrimitiveValue> consumeURL(CSSParserTokenRange& range)
+{
+    auto url = consumeURLRaw(range);
+    if (url.isNull())
+        return nullptr;
+    return CSSPrimitiveValue::createURI(url.toString());
+}
+
+} // namespace CSSPropertyParserHelpers
+} // namespace WebCore
