+++ Source/WebCore/css/parser/CSSPropertyParserHelpers.h
@@ -1,5 +1,6 @@
 // Copyright 2016 The Chromium Authors. All rights reserved.
 // Copyright (C) 2016-2023 Apple Inc. All rights reserved.
+// Copyright (C) 2024 Samuel Weinig <sam@webkit.org>
 //
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided that the following conditions are
@@ -51,7 +52,6 @@ namespace WebKitFontFamilyNames {
 enum class FamilyNamesIndex;
 }
 
-enum class BoxOrient : bool;
 enum class FontTechnology : uint8_t;
 
 // When these functions are successful, they will consume all the relevant
@@ -61,39 +61,6 @@ enum class FontTechnology : uint8_t;
 namespace CSSPropertyParserHelpers {
 
 RefPtr<CSSPrimitiveValue> consumeFontWeightNumber(CSSParserTokenRange&);
-
-RefPtr<CSSPrimitiveValue> consumeCustomIdent(CSSParserTokenRange&, bool shouldLowercase = false);
-RefPtr<CSSPrimitiveValue> consumeDashedIdent(CSSParserTokenRange&, bool shouldLowercase = false);
-
-StringView consumeStringRaw(CSSParserTokenRange&);
-RefPtr<CSSPrimitiveValue> consumeString(CSSParserTokenRange&);
-
-StringView consumeURLRaw(CSSParserTokenRange&);
-RefPtr<CSSPrimitiveValue> consumeURL(CSSParserTokenRange&);
-
-enum class PositionSyntax {
-    Position, // <position>
-    BackgroundPosition // <bg-position>
-};
-
-struct PositionCoordinates {
-    Ref<CSSValue> x;
-    Ref<CSSValue> y;
-};
-
-RefPtr<CSSValue> consumePosition(CSSParserTokenRange&, CSSParserMode, UnitlessQuirk, PositionSyntax);
-std::optional<PositionCoordinates> consumePositionCoordinates(CSSParserTokenRange&, CSSParserMode, UnitlessQuirk, PositionSyntax, NegativePercentagePolicy = NegativePercentagePolicy::Forbid);
-std::optional<PositionCoordinates> consumeOneOrTwoValuedPositionCoordinates(CSSParserTokenRange&, CSSParserMode, UnitlessQuirk);
-
-enum class AllowedImageType : uint8_t {
-    URLFunction = 1 << 0,
-    RawStringAsURL = 1 << 1,
-    ImageSet = 1 << 2,
-    GeneratedImage = 1 << 3
-};
-
-RefPtr<CSSValue> consumeImage(CSSParserTokenRange&, const CSSParserContext&, OptionSet<AllowedImageType> = { AllowedImageType::URLFunction, AllowedImageType::ImageSet, AllowedImageType::GeneratedImage });
-RefPtr<CSSValue> consumeImageOrNone(CSSParserTokenRange&, const CSSParserContext&);
 
 enum class AllowedFilterFunctions {
     PixelFilters,
@@ -189,8 +156,6 @@ RefPtr<CSSValue> consumeTranslate(CSSParserTokenRange&, CSSParserMode);
 RefPtr<CSSValue> consumeScale(CSSParserTokenRange&, CSSParserMode);
 RefPtr<CSSValue> consumeRotate(CSSParserTokenRange&, CSSParserMode);
 RefPtr<CSSValue> consumeRepeatStyle(CSSParserTokenRange&, const CSSParserContext&);
-RefPtr<CSSValue> consumePositionX(CSSParserTokenRange&, const CSSParserContext&);
-RefPtr<CSSValue> consumePositionY(CSSParserTokenRange&, const CSSParserContext&);
 RefPtr<CSSValue> consumePaintStroke(CSSParserTokenRange&, const CSSParserContext&);
 RefPtr<CSSValue> consumeListStyleType(CSSParserTokenRange&, const CSSParserContext&);
 RefPtr<CSSValue> consumePaintOrder(CSSParserTokenRange&);
