+++ Source/WebCore/css/parser/CSSPropertyParserWorkerSafe.cpp
@@ -2,6 +2,7 @@
 // Copyright (C) 2016-2024 Apple Inc. All rights reserved.
 // Copyright (C) 2021 Metrological Group B.V.
 // Copyright (C) 2021 Igalia S.L.
+// Copyright (C) 2024 Samuel Weinig <sam@webkit.org>
 //
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided that the following conditions are
@@ -46,6 +47,7 @@
 #include "CSSPropertyParserConsumer+List.h"
 #include "CSSPropertyParserConsumer+Number.h"
 #include "CSSPropertyParserConsumer+Percent.h"
+#include "CSSPropertyParserConsumer+URL.h"
 #include "CSSPropertyParserHelpers.h"
 #include "CSSToLengthConversionData.h"
 #include "CSSTokenizer.h"
